+++ src/Database/Esqueleto/Internal/PersistentImport.hs
@@ -138,6 +138,7 @@ module Database.Esqueleto.Internal.PersistentImport
   SqlType(..),
   UniqueDef(..),
   UpdateException(..),
+  -- * Here's a thing
   WhyNullable(..)
   ) where